backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==64.0.0
djangorestframework==3.16.1
idna==3.11
numpy==2.2.6
pgvector==0.4.1
psycopg==3.2.13
psycopg-binary==3.2.13
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
